LayoutTests/fast/mediastream/mediastreamtrack-video-torch.html

<!DOCTYPE html>
<html>
<head>
    <meta charset='utf-8'>
    <title>Camera track torch</title>
    <script src='../../resources/testharness.js'></script>
    <script src='../../resources/testharnessreport.js'></script>
</head>
<body>
    <canvas id='canvas'></canvas>
    <video controls autoplay width=640 height=480 playsInline id='video'></video>
    <script>

        let stream;
        promise_test(async (test) => {
            stream = await navigator.mediaDevices.getUserMedia({ video: { width : 640 } });
            assert_equals(stream.getVideoTracks()[0].getSettings().torch, undefined, 'settings-1');
            assert_equals(stream.getVideoTracks()[0].getCapabilities().torch, undefined, 'capabilities-1');

            stream = await navigator.mediaDevices.getUserMedia({ video: { width: 640, torch : true } });
            assert_equals(stream.getVideoTracks()[0].getCapabilities().torch, true, 'capabilities-2');

            await stream.getVideoTracks()[0].applyConstraints({ torch : true });
            assert_equals(stream.getVideoTracks()[0].getSettings().torch, true, 'settings-2');

            await stream.getVideoTracks()[0].applyConstraints({ torch : false });
            assert_equals(stream.getVideoTracks()[0].getSettings().torch, false, 'settings -3');

        }, 'getUserMedia torch');
    </script>
</body>
</html>
